backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
beautifulsoup4==4.12.3
selenium==4.25.0
webdriver-manager==4.0.2
httpx==0.27.2
pydantic==2.9.2
python-multipart==0.0.12
lxml==5.3.0
Pillow==10.4.0
cssutils==2.11.1
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
tqdm>=4.66.0
datasets>=2.20.0
huggingface-hub>=0.24.0
opencv-python-headless>=4.9.0
numpy>=1.24.0

# PDF Generation
reportlab>=4.0.0

pydantic-settings>=2.0.0
